重大修改：去除继承
修改为基于Role枚举的权限验证，并修改了路由中的装饰器验证。
