+++ app/api/charger.py
@@ -1,41 +1,42 @@
 from flask import Blueprint, request
 
-from .auth import loged_Veri
+from .auth import login_required
+
+from ..model import Role
 
 bp = Blueprint('charger', __name__, url_prefix='/charger')
 
 
 @bp.route('/', methods=('GET', 'POST'))
+@login_required(Role.CHARGE)
 def charger_index():
-    if loged_Veri(request.id) == 1:
-        return "This is staff index page"
+    pass
 
 
 @bp.route('/checkstaff', methods=('GET', 'POST'))
+@login_required(Role.CHARGE)
 def check_staff():
-    if loged_Veri(request.id) == 1:
-        return "This page will show the schedule of staff"
+    pass
 
 
 @bp.route('/arrangestaff', methods=('GET', 'POST'))
+@login_required(Role.CHARGE)
 def arrange_staff():
-    if loged_Veri(request.id) == 1:
-        return "This function will arrange timetable of staff"
+    pass
 
 
 @bp.route('/apppro', methods=('GET', 'POST'))
+@login_required(Role.CHARGE)
 def application_process():
-    if loged_Veri(request.id) == 1:
-        return "This is funcction will process the applicaiton of leave of staff "
+    pass
 
 
 @bp.route('/arrangeschedual', methods=('GET', 'POST'))
 def arrange_schedual():
-    if loged_Veri(request.id) == 1:
-        return "This url is for arrangeschedual"
+    pass
 
 
 @bp.route('/checkcktable', methods=('GET', 'POST'))
+@login_required(Role.CHARGE)
 def checktable():
-    if loged_Veri(request.id) == 1:
-        return "This url is for checktable"
+    pass
