+++ app/api/manager.py
@@ -1,44 +1,15 @@
 from flask import Blueprint, json, jsonify, request
 
-from ..model import Manager
-from .auth import loged_Veri
+from .util import login_required
 
 bp = Blueprint('manager', __name__, url_prefix='/manager')
 
 
 @bp.route('/checkstaff', methods=['GET'])
 def check_staff():
-    request_data = json.loads(str(request.get_data(), 'utf-8'))
-    if loged_Veri(request_data['UserId']) == 2:
-        manager = Manager(request_data['UserId'])
-        response_data = manager.retrieve_employee()
-        response_data = {
-            'status': 20000,
-            'data': response_data
-        }
-    else:
-        response_data = {
-            'status': 50000,
-            'data': {}
-        }
-    response_data = jsonify(response_data)
-    return response_data
+    pass
 
 
 @bp.route('/updatestaff', methods=['PUT'])
 def update_staff():
-    request_data = json.loads(str(request.get_data(), 'utf-8'))
-    if loged_Veri(request_data['UserId']) == 2:
-        manager = Manager(request_data['UserId'])
-        manager.update_employee(request_data['data'])
-        response_data = {
-            'status': 20000,
-            'data': {}
-        }
-    else:
-        response_data = {
-            'status': 50000,
-            'data': {}
-        }
-    response_data = jsonify(response_data)
-    return response_data
+    pass
