+++ app/model/User.py
@@ -1,4 +1,5 @@
 import os
+from enum import IntEnum
 
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask import Flask
@@ -6,6 +7,16 @@ from flask_sqlalchemy import SQLAlchemy
 
 from .. import db
 from .Department import Department
+
+
+class Role(IntEnum):
+    """职务
+
+    使用数值枚举类，该类是int的子类
+    """
+    STAFF = 1
+    CHARGE = 2
+    MANAGER = 3
 
 
 class User(db.Model):
@@ -39,7 +50,7 @@ class User(db.Model):
 
     def __repr__(self):
         """打印对象的信息"""
-        return '<User {}, ID={}>'.format(self.name, self.ID)
+        return '<User {}, ID={}, Role={}>'.format(self.name, self.ID, Role(self.identity).name)
 
     @property
     def password(self):
@@ -51,6 +62,14 @@ class User(db.Model):
 
     def verify_password(self, password):
         return check_password_hash(self.password_hash, password)
+
+    @property
+    def role(self):
+        return Role(self.identity)
+
+    @role.setter
+    def role(self, r):
+        self.identity = r
 
     #
     # 数据库查询方法
@@ -64,7 +83,6 @@ class User(db.Model):
         """
         return User.query.all()
 
-
     @staticmethod
     def ByID(ID):
         """根据ID构造对象
@@ -74,7 +92,7 @@ class User(db.Model):
         return User.query.get(ID)
 
     #
-    # 自修改方法
+    # 公共方法
     #
 
     def login(self, password):
@@ -100,3 +118,69 @@ class User(db.Model):
         """
         db.session.delete(self)
         db.session.commit()
+
+    #
+    # 主管方法
+    #
+
+    def arrange_work(self):
+        """安排工作班次"""
+        pass
+
+    def update_work_arrangement(self):
+        """修改本部门员工工作安排"""
+        pass
+
+    def approve_leave(self):
+        """请假审批"""
+        pass
+
+    def approve_report(self):
+        """销假处理"""
+        pass
+
+    def approve_overtime(self):
+        """加班审批"""
+        pass
+
+    #
+    # 经理方法
+    #
+
+    def update_employee(self, u):
+        """修改员工信息
+
+        :param u : User
+        """
+        employee.user.update_self()
+
+    def retrieve_employee(self, employeeID):
+        """查看员工信息"""
+        return
+
+    def get_employees(self):
+        """查看全部员工的信息
+
+        :return List[Employee]
+        """
+        return
+
+    def delete_employee(self, employee):
+        """人员删除"""
+        employee.delete_self()
+
+    def update_position(self):
+        """身份修改"""
+        pass
+
+    def create_employee(self):
+        """人员添加"""
+        pass
+
+    def release_temporary_overtime(self):
+        """发布全单位加班"""
+        pass
+
+    def retrieve_work_situation(self):
+        """查看上班情况"""
+        pass
