+++ app/model/__init__.py
@@ -1,10 +1,10 @@
-from .Charge import Charge
+
 from .Department import Department
-from .Employee import Employee
+
 from .Leave import Leave
-from .Manager import Manager
+
 from .Overtime import Overtime, OvertimeSituation, TemporaryOvertime
 from .Report import Report
 from .Signsheet import Punch, SignSheet, WorkSituation
-from .User import User
+from .User import User, Role
 from .WorkArrangement import WorkArrangement
